Implement grid reflowing on resize

+++ src/pane.c
@@ -18,14 +18,12 @@
 extern pid_t forkpty(int *, char *, struct termios *, struct winsize *);
 
 void pane_destroy(struct pane *pane) {
-  if (pane->pty)
-    close(pane->pty);
+  if (pane->pty) close(pane->pty);
   if (pane->pid) {
     int status;
     kill(pane->pid, SIGTERM);
     pid_t result = waitpid(pane->pid, &status, WNOHANG);
-    if (result == -1)
-      die("waitpid:");
+    if (result == -1) die("waitpid:");
   }
   pane->pty = 0;
   pane->pid = 0;
@@ -36,29 +34,29 @@ void pane_destroy(struct pane *pane) {
 
 struct pane *pane_from_pty(struct pane *p, int pty) {
   for (; p; p = p->next)
-    if (p->pty == pty)
-      return p;
+    if (p->pty == pty) return p;
   return nullptr;
 }
 
 void pane_draw(struct pane *pane, bool redraw, struct string *outbuffer) {
   static uint8_t fg, bg;
   static uint32_t attr;
-  struct grid *g = pane->fsm.opts.alternate_screen ? &pane->fsm.alternate
-                                                   : &pane->fsm.primary;
+  struct grid *g = pane->fsm.opts.alternate_screen ? &pane->fsm.alternate : &pane->fsm.primary;
   char fmt[100];
   for (int i0 = 0; i0 < g->h; i0++) {
     int row = (i0 + g->offset) % g->h;
-    if (!redraw && !g->dirty[row])
-      continue;
+    if (!redraw && !g->dirty[row]) continue;
     g->dirty[row] = false;
     int lineno = 1 + pane->y + i0;
     int columnno = 1 + pane->x;
+    struct cell *line = &g->cells[row * g->w];
+    int line_length = MIN(line->n_significant, g->w);
+    // TODO: Get rid of snprintf since it's really slow
     int n = snprintf(fmt, sizeof(fmt), "\x1b[%d;%dH", lineno, columnno);
     string_push(outbuffer, fmt, n);
 
-    for (int j = 0; j < g->w; j++) {
-      struct cell *c = &g->cells[row * g->w + j];
+    for (int col = 0; col < line_length; col++) {
+      struct cell *c = &line[col];
       if (c->fg != fg) {
         // TODO: apply fg
         fg = c->fg;
@@ -71,8 +69,12 @@ void pane_draw(struct pane *pane, bool redraw, struct string *outbuffer) {
         // TODO: apply attributes
         attr = c->attr;
       }
-      string_push(outbuffer, (char*)c->symbol.utf8, c->symbol.len);
+      string_push(outbuffer, (char *)c->symbol.utf8, c->symbol.len);
     }
+
+    int num_blanks = g->w - line_length;
+    if (num_blanks > 0)
+      string_memset(outbuffer, ' ', num_blanks);
   }
 }
 
@@ -89,10 +91,8 @@ void pane_resize(struct pane *pane, int w, int h) {
   // if (debugsize) w = h = debugsize;
   if (pane->w != w || pane->h != h) {
     struct winsize ws = {.ws_col = w, .ws_row = h};
-    if (pane->pty)
-      ioctl(pane->pty, TIOCSWINSZ, &ws);
-    if (pane->pid)
-      kill(pane->pid, SIGWINCH);
+    if (pane->pty) ioctl(pane->pty, TIOCSWINSZ, &ws);
+    if (pane->pid) kill(pane->pid, SIGWINCH);
     pane->w = w;
     pane->h = h;
   }
@@ -101,8 +101,7 @@ void pane_resize(struct pane *pane, int w, int h) {
 void pane_start(struct pane *pane) {
   struct winsize panesize = {.ws_col = pane->w, .ws_row = pane->h};
   pid_t pid = forkpty(&pane->pty, NULL, NULL, &panesize);
-  if (pid < 0)
-    die("forkpty:");
+  if (pid < 0) die("forkpty:");
 
   if (pid == 0) {
     execlp(pane->process, pane->process, NULL);
@@ -135,7 +134,6 @@ void pane_remove(struct pane **lst, struct pane *rem) {
 
 int pane_count(struct pane *pane) {
   int n = 0;
-  for (; pane; pane = pane->next)
-    n++;
+  for (; pane; pane = pane->next) n++;
   return n;
 }
